+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.9
 **Bugfixes**
 - Fixed crash with Grid
+- Priority field can now display 4 digits at a time
 
 **Features**
 - Added fluid storage